setup/requirements.txt
pandas==1.5.2
grpcio==1.47.0
prefect==2.8.5
prefect-sqlalchemy==0.2.2
prefect-gcp[cloud_storage]==0.3.0
protobuf==4.21.11
pyarrow==10.0.1
psycopg2-binary==2.9.5
sqlalchemy==1.4.46
dbt-bigquery==1.4.3
dbt-core==1.4.5
requests==2.28.2
prefect-dbt[cli]==0.3.1
prefect-dbt[bigquery]==0.3.1
python-dotenv==1.0.0